MOve code responsible for grabbing a highlighted chunk of text down to the FindUsages layer.
